Validation and Exceptions

+++ src/main/java/com/agh/surveys/service/group/GroupService.java
@@ -1,10 +1,8 @@
 package com.agh.surveys.service.group;
 
 
+import com.agh.surveys.exception.BusinessException;
 import com.agh.surveys.exception.group.GroupNotFoundException;
-import com.agh.surveys.exception.group.UserAlreadyInTheGroupException;
-import com.agh.surveys.exception.group.UserNotInTheGroupException;
-import com.agh.surveys.exception.user.UserNotFoundException;
 import com.agh.surveys.model.group.Group;
 import com.agh.surveys.model.group.dto.GroupCreateDto;
 import com.agh.surveys.model.group.dto.GroupRespDto;
@@ -22,7 +20,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
-import java.util.Collections;
+import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -48,34 +46,55 @@ public class GroupService implements IGroupService {
     @Autowired
     QuestionService questionService;
 
+    private static final int pollDeadlineMarginMinutes = 2;
+
     @Override
     public Poll addPolltoGroup(PollCreateDto pollCreateDto, Integer groupId) {
-        Group group = groupRepository.findById(groupId).orElseThrow(GroupNotFoundException::new);
-        User author = userService.getUserByNick(pollCreateDto.getAuthorId());
+
+        Group group = getGroup(groupId);
+        User author = userService.getUserByNick(pollCreateDto.getAuthorNick());
+
+        validatePollDto(pollCreateDto);
+
+        if (!userService.isUserInGroup(author, group)) {
+            throw new BusinessException("Poll author must be member of a group.");
+        }
+
         List<Question> questions = new LinkedList<>();
         Poll poll = new Poll(pollCreateDto.getPollName(), LocalDateTime.now(), pollCreateDto.getPolDeadline(), author, questions);
         pollService.savePoll(poll);
         questions.addAll(questionService.addAllQuestionDetails(poll, pollCreateDto.getQuestionDetails()));
-        pollService.savePoll(poll);
-        group.getGroupPolls().add(poll);
-        groupRepository.save(group);
+        poll.setPollGroup(group);
+
         return poll;
     }
 
-    @Override
-    public Integer addGroup(GroupCreateDto groupCreateDto) {
-        User groupLeader = userRepository.getOne(groupCreateDto.getLeaderNick());
-        List<User> members = Collections.emptyList(); // TODO dla pustej listy membersNick jest wyrzucany wyjątek
-        if ( !groupCreateDto.getGroupMembersNicks().isEmpty()) {
-            for (String userName : groupCreateDto.getGroupMembersNicks()) {
-                User user = userService.getUserByNick(userName);
-                members.add(user);
-            }
+    private LocalDateTime createdPollDeadlineMargin(){
+        return LocalDateTime.now().plusMinutes(pollDeadlineMarginMinutes);
+    }
+
+    private void validatePollDto(PollCreateDto pollCreateDto){
+        if(pollCreateDto.getPolDeadline().isBefore(createdPollDeadlineMargin()) ){
+            throw new BusinessException("Created poll can't have deadline in the past.");
         }
+    }
+
+    @Override
+    public GroupRespDto addGroup(GroupCreateDto groupCreateDto) {
+        User groupLeader = userRepository.getOne(groupCreateDto.getLeaderNick());
+        List<User> members = new ArrayList<>(); // TODO dla pustej listy membersNick jest wyrzucany wyjątek
+
+        for (String userName : groupCreateDto.getGroupMembersNicks()) {
+            User user = userService.getUserByNick(userName);
+            members.add(user);
+        }
+        //lider nie moze sie duplikować wsrod czlonkow
+        members.remove(groupLeader);
+
         Group group = new Group(groupCreateDto.getGroupName(), groupLeader, members);
+        groupRepository.save(group);
 
-        return groupRepository.save(group).getId();
-
+        return new GroupRespDto(group);
     }
 
     @Override
@@ -85,53 +104,51 @@ public class GroupService implements IGroupService {
 
     @Override
     public void removeGroup(Integer id) {
-        groupRepository.deleteById(id);
+        if (groupRepository.existsById(id)){
+            groupRepository.deleteById(id);
+        }else{
+            throw new GroupNotFoundException();
+        }
     }
 
     @Override
     public GroupRespDto getGroupDto(Integer id) {
-        Group group=  groupRepository.findById(id)
+        Group group = groupRepository.findById(id)
                 .orElseThrow(GroupNotFoundException::new);
         return new GroupRespDto(group);
     }
 
     @Override
-    public Group getGroup(Integer id){
+    public Group getGroup(Integer id) {
         return groupRepository.findById(id).orElseThrow(GroupNotFoundException::new);
     }
 
-    //@TODO Maybe there is some way to update only group and the user will be updated as well? (LK)
     @Override
     public void addGroupMember(Integer groupId, String userNick) {
         Group group = groupRepository.getOne(groupId);
         User user = userRepository.getOne(userNick);
 
-        if(group.getGroupMembers().contains(user)){
-            throw new UserAlreadyInTheGroupException();
+        if (group.getGroupMembers().contains(user)) {
+            throw new BusinessException("This user is already in the group.");
         }
 
         group.getGroupMembers().add(user);
 
-        saveGroupAndUser(group, user);
-    }
-
-    private void saveGroupAndUser(Group group, User user) {
         groupRepository.save(group);
-        userRepository.save(user);
     }
 
     @Override
     public void removeGroupMember(Integer groupId, String userNick) {
-        Group group = groupRepository.getOne(groupId);
-        User user = userRepository.getOne(userNick);
+        Group group = getGroup(groupId);
+        User user = userService.getUserByNick(userNick);
 
-        if(!group.getGroupMembers.contains(user)){
-            throw new UserNotInTheGroupException();
+        if (!group.getGroupMembers().contains(user)) {
+            throw new BusinessException("Provided user is not in this group.");
         }
 
         group.getGroupMembers().remove(user);
 
-        saveGroupAndUser(group, user);
+        groupRepository.save(group);
 
     }
 }
